# api/gen/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: skullandbonesdao@gmail.com
  description: This is the Skull&Bones Trades API V2
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: S&B - StarAtlas Trades API
  version: 1.0.0
servers:
- description: API V2 Endpoint
  url: https://api.skullnbones.xyz/v2
- description: API V2 Endpoint
  url: https://localhost/v2
tags:
- description: UDF Compatible API Calls
  name: UDF
paths:
  /:
    get:
      description: |
        Gets a status message
      operationId: getStatus
      responses:
        "200":
          content:
            text/plain:
              schema:
                example: Weclome to S&B StarAtlas Trades API
                type: string
                x-content-type: text/plain
          description: status message resutlt
        "400":
          description: error
        "404":
          description: not found
      summary: status
      tags:
      - UDF
      x-swagger-router-controller: UDF
      x-eov-operation-handler: controllers/UDFController
  /time:
    get:
      description: |
        Gets server time
      operationId: getTime
      responses:
        "200":
          content:
            text/plain:
              schema:
                example: 1662761714
                type: number
                x-content-type: text/plain
          description: time message resutlt
        "400":
          description: error
        "404":
          description: not found
      summary: time
      tags:
      - UDF
      x-swagger-router-controller: UDF
      x-eov-operation-handler: controllers/UDFController
  /config:
    get:
      description: |
        Gets a config message
      operationId: getConfig
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Config'
          description: search results matching criteria
        "400":
          description: error
        "404":
          description: not found
      summary: src
      tags:
      - UDF
      x-swagger-router-controller: UDF
      x-eov-operation-handler: controllers/UDFController
  /symbol_info:
    get:
      description: |
        Gets a symbols info message
      operationId: geSymbolInfo
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SymbolInfo'
          description: returns symbols info
        "400":
          description: error
        "404":
          description: not found
      summary: symbol into
      tags:
      - UDF
      x-swagger-router-controller: UDF
      x-eov-operation-handler: controllers/UDFController
  /symbols:
    get:
      description: |
        Gets a symbol info message
      operationId: geSymbol
      parameters:
      - description: trading pair symbol
        explode: true
        in: query
        name: symbol
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Symbol'
          description: returns symbol info
        "400":
          description: error
        "404":
          description: not found
      summary: symbol into
      tags:
      - UDF
      x-swagger-router-controller: UDF
      x-eov-operation-handler: controllers/UDFController
  /search:
    get:
      description: |
        Gets a symbol info message
      operationId: geSearchSymbol
      parameters:
      - description: trading pair symbol
        explode: true
        in: query
        name: query
        required: false
        schema:
          type: string
        style: form
      - description: limit results
        explode: true
        in: query
        name: limit
        required: false
        schema:
          type: string
        style: form
      - description: type
        explode: true
        in: query
        name: type
        required: false
        schema:
          type: string
        style: form
      - description: exchange
        explode: true
        in: query
        name: exchange
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SearchSymbol'
                type: array
                x-content-type: application/json
          description: returns symbol info
        "400":
          description: error
        "404":
          description: not found
      summary: symbol into
      tags:
      - UDF
      x-swagger-router-controller: UDF
      x-eov-operation-handler: controllers/UDFController
  /history:
    get:
      description: |
        Gets a symbol info message
      operationId: geHistoryElement
      parameters:
      - description: trading pair symbol
        explode: true
        in: query
        name: symbol
        required: false
        schema:
          type: string
        style: form
      - description: timestamp from
        explode: true
        in: query
        name: from
        required: false
        schema:
          type: number
        style: form
      - description: timestamp to
        explode: true
        in: query
        name: to
        required: false
        schema:
          type: number
        style: form
      - description: time resolution
        explode: true
        in: query
        name: resolution
        required: false
        schema:
          type: string
        style: form
      - description: counter
        explode: true
        in: query
        name: countback
        required: false
        schema:
          type: number
        style: form
      - description: symbol-cde
        explode: true
        in: query
        name: currencyCode
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryElement'
          description: returns symbol info
        "400":
          description: error
        "404":
          description: not found
      summary: history object
      tags:
      - UDF
      x-swagger-router-controller: UDF
      x-eov-operation-handler: controllers/UDFController
components:
  schemas:
    Config:
      example:
        supports_marks: false
        supports_time: true
        exchanges:
        - name: StarAtlas
          value: SA
          desc: StarAtlas Exch
        - name: StarAtlas
          value: SA
          desc: StarAtlas Exch
        supports_search: true
        supports_group_request: false
        symbols_types:
        - name: Cryptocurrency
          value: crypto
        - name: Cryptocurrency
          value: crypto
        supported_resolutions:
        - 1D
        - 1D
        homePage: https://www.acme-corp.com
        supports_timescale_marks: false
      properties:
        exchanges:
          items:
            $ref: '#/components/schemas/Exchange'
          type: array
        symbols_types:
          items:
            $ref: '#/components/schemas/SymbolsType'
          type: array
        supported_resolutions:
          items:
            example: 1D
            type: string
          type: array
        homePage:
          example: https://www.acme-corp.com
          format: url
          type: string
        supports_search:
          example: true
          type: boolean
        supports_group_request:
          example: false
          type: boolean
        supports_marks:
          example: false
          type: boolean
        supports_timescale_marks:
          example: false
          type: boolean
        supports_time:
          example: true
          type: boolean
      required:
      - exchnages
      type: object
    SymbolInfo:
      example:
        symbol:
        - PX4ATLAS
        - PX4ATLAS
        ticker:
        - PX4ATLAS
        - PX4ATLAS
        has_daily: true
        has_weekly_and_monthly: 0.8008281904610115
        minmov2: 0
        session: 24x7
        timezone: UTC
        has_intraday: true
        description:
        - PX4 ATLAS
        - PX4 ATLAS
        listed_exchange: SA
        supported_resolutions:
        - "1"
        - "1"
        type: crypto
        currency_code:
        - BTC
        - BTC
        full_name:
        - PX4ATLAS
        - PX4ATLAS
        name:
        - PX4ATLAS
        - PX4ATLAS
        minmovement: 1
        minmov: 1
        exchange: SA
        pricescale:
        - 100
        - 100
        minmovement2: 0
        data_status: streaming
      properties:
        symbol:
          items:
            example: PX4ATLAS
            type: string
          type: array
        ticker:
          items:
            example: PX4ATLAS
            type: string
          type: array
        name:
          items:
            example: PX4ATLAS
            type: string
          type: array
        full_name:
          items:
            example: PX4ATLAS
            type: string
          type: array
        description:
          items:
            example: PX4 ATLAS
            type: string
          type: array
        exchange:
          example: SA
          type: string
        listed_exchange:
          example: SA
          type: string
        type:
          example: crypto
          type: string
        currency_code:
          items:
            example: BTC
            type: string
          type: array
        session:
          example: 24x7
          type: string
        timezone:
          example: UTC
          type: string
        minmovement:
          example: 1
          type: number
        minmov:
          example: 1
          type: number
        minmovement2:
          example: 0
          type: number
        minmov2:
          example: 0
          type: number
        pricescale:
          items:
            example: 100
            type: number
          type: array
        supported_resolutions:
          items:
            example: "1"
            type: string
          type: array
        has_intraday:
          example: true
          type: boolean
        has_daily:
          example: true
          type: boolean
        has_weekly_and_monthly:
          type: number
        data_status:
          example: streaming
          type: string
      type: object
    Symbol:
      example:
        symbol: PX4ATLAS
        ticker: PX4ATLAS
        has_daily: true
        has_weekly_and_monthly: 0.8008281904610115
        minmov2: 0
        session: 24x7
        timezone: UTC
        has_intraday: true
        description: PX4 ATLAS
        listed_exchange: SA
        supported_resolutions:
        - "1"
        - "1"
        type: crypto
        currency_code: BTC
        full_name: PX4ATLAS
        name: PX4ATLAS
        minmovement: 1
        minmov: 1
        exchange: SA
        pricescale:
        - 100
        - 100
        minmovement2: 0
        data_status: streaming
      properties:
        symbol:
          example: PX4ATLAS
          type: string
        ticker:
          example: PX4ATLAS
          type: string
        name:
          example: PX4ATLAS
          type: string
        full_name:
          example: PX4ATLAS
          type: string
        description:
          example: PX4 ATLAS
          type: string
        exchange:
          example: SA
          type: string
        listed_exchange:
          example: SA
          type: string
        type:
          example: crypto
          type: string
        currency_code:
          example: BTC
          type: string
        session:
          example: 24x7
          type: string
        timezone:
          example: UTC
          type: string
        minmovement:
          example: 1
          type: number
        minmov:
          example: 1
          type: number
        minmovement2:
          example: 0
          type: number
        minmov2:
          example: 0
          type: number
        pricescale:
          items:
            example: 100
            type: number
          type: array
        supported_resolutions:
          items:
            example: "1"
            type: string
          type: array
        has_intraday:
          example: true
          type: boolean
        has_daily:
          example: true
          type: boolean
        has_weekly_and_monthly:
          type: number
        data_status:
          example: streaming
          type: string
      type: object
    SearchSymbol:
      example:
        symbol: PX4ATLAS
        ticker: PX4ATLAS
        full_name: PX4ATLAS
        description: PX4 ATLAS
        exchange: SA
        type: crypto
      properties:
        symbol:
          example: PX4ATLAS
          type: string
        full_name:
          example: PX4ATLAS
          type: string
        description:
          example: PX4 ATLAS
          type: string
        exchange:
          example: SA
          type: string
        ticker:
          example: PX4ATLAS
          type: string
        type:
          example: crypto
          type: string
      type: object
    HistoryElement:
      example:
        s: ok
        c:
        - 12322121
        - 12322121
        t:
        - 12322121
        - 12322121
        v:
        - 12322121
        - 12322121
        h:
        - 12322121
        - 12322121
        l:
        - 12322121
        - 12322121
        o:
        - 12322121
        - 12322121
      properties:
        s:
          example: ok
          type: string
        t:
          items:
            example: 12322121
            type: number
          type: array
        c:
          items:
            example: 12322121
            type: number
          type: array
        o:
          items:
            example: 12322121
            type: number
          type: array
        h:
          items:
            example: 12322121
            type: number
          type: array
        l:
          items:
            example: 12322121
            type: number
          type: array
        v:
          items:
            example: 12322121
            type: number
          type: array
      type: object
    Exchange:
      example:
        name: StarAtlas
        value: SA
        desc: StarAtlas Exch
      properties:
        value:
          example: SA
          type: string
        name:
          example: StarAtlas
          type: string
        desc:
          example: StarAtlas Exch
          type: string
      required:
      - exchange
      type: object
    SymbolsType:
      example:
        name: Cryptocurrency
        value: crypto
      properties:
        value:
          example: crypto
          type: string
        name:
          example: Cryptocurrency
          type: string
      type: object